perf: refactor copmonents packages
