+++ components/atoms/src/index.ts
@@ -1,14 +1,13 @@
-export { Button } from "./Button";
-export type { ButtonProps } from "./Button";
-export { ButtonLink } from "./ButtonLink";
-export { Input } from "./Input";
-export { Select } from "./Select";
-export { Toast } from "./Toast";
-export { Image } from "./Image";
-export { Avatar } from "./Avatar";
-export { Switch } from "./Switch";
-export { Checkbox } from "./Checkbox";
-export { Spinner } from "./Spinner";
-export { Link } from "./Link";
-export { Badge } from "./Badge";
-export { Text } from "./Text";
+export * from "./Button";
+export * from "./ButtonLink";
+export * from "./Input";
+export * from "./Select";
+export * from "./Toast";
+export * from "./Image";
+export * from "./Avatar";
+export * from "./Switch";
+export * from "./Checkbox";
+export * from "./Spinner";
+export * from "./Link";
+export * from "./Badge";
+export * from "./Text";
