+++ components/molecules/src/index.ts
@@ -1,5 +1,5 @@
-export { ButtonGroup } from "./ButtonGroup";
-export { CircleIndicator } from "./CircleIndicator";
-export { GlowLogo } from "./GlowLogo";
-export { Parallax } from "./Parallax";
-export { ModalScreen } from "./ModalScreen";
+export * from "./ButtonGroup";
+export * from "./CircleIndicator";
+export * from "./GlowLogo";
+export * from "./Parallax";
+export * from "./ModalScreen";
