+++ components/organisms/src/ButtonCTA/index.ts
@@ -1,1 +1,2 @@
 export { ButtonCTA } from "./ButtonCTA";
+export * from "./styles";
