+++ components/organisms/src/ButtonCTA/styles/index.ts
@@ -1,0 +1,2 @@
+export { DaoboardButtonCTA } from "./DaoboardButtonCTA";
+export { MoshButtonCTA } from "./MoshButtonCTA";
