+++ components/screens/src/index.ts
@@ -1,4 +1,4 @@
-export { ButtonScreen } from "./ButtonScreen";
-export { LandingDaoboardScreen } from "./LandingDaoboardScreen";
-export { MetafamLandingScreen } from "./MetafamLandingScreen";
-export { LandingMoshScreen } from "./LandingMoshScreen";
+export * from "./ButtonScreen";
+export * from "./LandingDaoboardScreen";
+export * from "./MetafamLandingScreen";
+export * from "./LandingMoshScreen";
